Move nutrition summary bar to the bottom
