+++ src/frontend/src/widgets/NutritionSummaryWidget/index.ts
@@ -1,1 +1,2 @@
 export * from './NutritionSummaryWidget';
+export * from './NutritionSummaryWidgetBar';
